Added an ErrorResponse object to standardize error responses like our API Playbook suggests.
Since we never had this before, I figured it's for the best to add enough constructor overloads so that the use of this object wouldn't be a hindrance. When someone else takes over, they're more likely to follow the standard if it's easier to.

Also added tests as some of the beheviour is slightly more complicated than your typical constructor.