Implemented GetTenancyTransactionsValidator.
Added the typical "not null", "not empty" checks.
Added postcode validation - taken the code from LBHAddressesAPI as I already TDD'ed this same thing before.

That raises an interesting question though. Shouldn't we have some sort of 'Validations' project that's TDD'ed. The idea is that we always know that that code is alright and tested, so we can simply bring it into other projects without the tests.

Also a TODO.. our security expert says we need a certain security check, that pretty much requires data access from validator. So I'm leaving this out until I get to that point.