Extract out getting TenancyDetails into its own endpoint.
After some consideration there was an agreement within the team that tenancy details coming together with the transactions are too painful to look at.

Hence the code refactoring to extract out the tenancy details into their own endpoint, which would be much easier to move out into a different api in the future.

While doing this I had to also change the Route attributes or else there would be a url conflict between endpoints.

Squashed commit message #2:
Added missing parentheses to the usecase.