Added a 404 response to GetTenancyDetails endpoint.
Since this is a single resource, as per standard when it's not found, it needs to return a 404 status code (unlike empty collections).

Added an error message to show what request data was sent. However, I think this response object should contain the request object just like with the success case. After all having a request object back is most useful upon failure, not the success.