Added GetTenancyDetails request validator.
Just like with the GetAllTenancyTransactions endpoint, the post code validation needs to be changed to only accept full post code - not the partial one like it does now.

This is a copy paste of the other validator. I decided to have as a separate class as this GetTenancyDetails mehtod is likely to be moved out into a separate API along with the validator.