Untied the GetAllTenancyTransactions endpoint from GetTenancyDetails endpoint.
Did this by inserting an extra query to get the desired information rather than calling an extra endpoint that returns too much information.

Ideally, this extra query would be combined with the main one. But no time for that.

Extra Query:  "TransactionsGateway.cs Ln:160-Ln:171"

Also had to wrap up all the old functinoality of "GetAllTenancyTransactionStatements method" into an IF block (Ln:175) to deal with the case where the tag_ref is not found.