Changed the postcode validation regex to only allow full post code.
Also removed couple nuances of some characters not being used at certain positions. Because if I did that, the tests would explode, the regex would be massive - and there was no request to check whether the postcode is valid. So I made it so it only checks the format of it.