Modified the regex for TenancyDetails request validator.
Didn't add any tests, as I'm using the same regex as the one that's already tested. Now, I have duplicate functionality and repeated code, which could be resolved using abstract class.

However!!!! It's not the issue of code - it's an issue of standard we're using. As per our current standard the boundary objects get named according to the endpoint they belong in.

Issue with that is that if you have 2 endpoints accepting the same input, you  now need to pointlessly have to different request boundary objects, just because our naming standard requires request object be named after the endpoint. And it doesn't make sense to reuse the one from the other endpoint due to its name being related to that endpoint!

Now what's the issue in all of that? Well, when creating validator - you create validator per request object due to how FluentValidations work. So you also end up with repeated code for the validator.

This doesn't seem that bad when you think about that request object and validator don't contain much code... but the tests.... when you consider tests it all goes really bad. Because now you have to have the same tests running twice - taking twice the time to run them in addition to having that same massive tests code twice.

This needs to change. Need a smarter way to name these objects.

And as I wrote before, even though this could be resolved with abstract class, I'm unsure if we should be okay with havinf duplicated request objects.