Reverted CI/CD pipeline to old, modified to trigger on this temp branch.
Due to the infrastructure problems on the environment related to new pipeline, the old pipeline will be used as a temporary measure to get the service out there. I updated the pipeline to fire off on this branch rather than waiting to be merged to master (as we don't want to use this pipeline apart from this occasion (and maybe one more). Also changed environment variable names as I don't want to overwrite the current ones for the new pipeline on CCI project configuration.