Added PostCodeBaseValidator.
This is the 1st step at getting rid of duplicate validator tests accross an API. While here on transactions it would be less visible (due to them only being repeated twice), on the previous API I worked on exact same tests have been repeated as much as 4 times.

The change to an architecture I though of involves dependecy injection into the main validator. And building up of that validator like a lego from base validators - so that the functionality that belongs within those base valiators would only be needed to test once!

So this is a base validator for post codes... I think ideally we would group them based of the type. So for instance, when it comes to postcodes, you'd have all the post codes related validation like 'partial post code format', 'full post code format', 'partial postcode characters', 'full postcode characters' (depending  on which ones of these your API needs).