Added tests for GetTenancyDetailsValidator, and changed it to new Validator architecture.
Also added a test helper method to create the request object. Now, that is indeed duplicate code, however, there's no guarantee that these endpoints would stay the same when it comes to accepting parameters, hence it's for the better to keep them decoupled like that.

Also... the Tenancy Details endpoint is likely to be taken away into separate API at some point. So having every bit of it decoupled (even if code is duplicated) is actually reasonable, as it will mean less work when moving all of it out.